+++ .gitignore
@@ -1,3 +1,1 @@
 /target
-test.dbg
-Cargo.lock
