+++ anita_core/.gitignore
@@ -1,0 +1,3 @@
+/target
+test.dbg
+Cargo.lock
