+++ anita_core/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "anita_core"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anita_macros = { path = "../anita_macros" }
+cranelift = "0.114.0"
+cranelift-module = "0.114.0"
+cranelift-jit = {version = "0.114.0", optional = true}
+cranelift-native = "0.114.0"
+evalexpr = "12.0.1"
+itertools = "0.13.0"
+peg = "0.8.4"
+
+[features]
+default = ["cranelift-jit"]
+selinux-fix = ["cranelift-jit/selinux-fix"]
+
+[target.'cfg(target_os = "linux")'.dependencies]
+cranelift-jit = {version = "0.114.0", features = ["selinux-fix"]}
