+++ anita_core/src/jit/mod.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use super::function_manager::{DefaultFunctionManager, FunctionManager};
+use super::function_manager::FunctionManager;
 use codegen::ir::FuncRef;
 use cranelift::{
     codegen,
@@ -29,9 +29,9 @@ macro_rules! compile_expression {
         {
             use std::mem;
             use $crate::jit::{compiled_function::CompiledFunction, JIT};
-            use $crate::function_manager::DefaultFunctionManager;
+            use $crate::function_manager::NoFunctions;
 
-            let mut jit = JIT::<f32, DefaultFunctionManager>::default();
+            let mut jit = JIT::<f32, NoFunctions>::default();
             match jit.compile($expression, &[$( stringify!($parameter) ),*]) {
                 Ok(code_ptr) => {
                     let function_pointer = unsafe { mem::transmute::<*const u8, fn($(compile_expression!(@to_f32 $parameter)),+) -> f32>(code_ptr) };
@@ -45,12 +45,12 @@ macro_rules! compile_expression {
         }
     };
 
-    ($expression:expr, ($($parameter:ident),+) -> f32, $function_manager:ty) => {
+    ($expression:expr, ($($parameter:ident),+) -> f32, $functions:ty) => {
         {
             use std::mem;
             use $crate::jit::{compiled_function::CompiledFunction, JIT};
 
-            let mut jit = JIT::<f32, $function_manager>::default();
+            let mut jit = JIT::<f32, $functions>::default();
             match jit.compile($expression, &[$( stringify!($parameter) ),*]) {
                 Ok(code_ptr) => {
                     let function_pointer = unsafe { mem::transmute::<*const u8, fn($(compile_expression!(@to_f32 $parameter)),+) -> f32>(code_ptr) };
@@ -92,7 +92,7 @@ impl From<ParseError<LineCol>> for JITError {
     }
 }
 
-pub struct JIT<T: AnitaType, F: FunctionManager = DefaultFunctionManager> {
+pub struct JIT<T: AnitaType, F: FunctionManager> {
     builder_context: FunctionBuilderContext,
     ctx: codegen::Context,
     module: Box<JITModule>,
