+++ anita_macros/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "internal_macros"
+name = "anita_macros"
 version = "0.1.0"
 edition = "2021"
