--- integration_tests/.gitignore
@@ -1,1 +1,0 @@
-/target
