--- integration_tests/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "integration_tests"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anita ={ path = "../", optional = true }
-
-[features]
-default = ["anita"]
-no-default-functions = ["anita/no-default-functions"]
